requirements.txt
redis
argh
asn1crypto
bcrypt
certifi
cffi
chardet
cryptography
decorator
idna
Jinja2
jsonpath-rw
jsonpath-rw-ext
MarkupSafe
paramiko
pathtools
pbr
ply
pyasn1
pycparser
PyMySQL
PyNaCl
PyYAML
requests
selenium
six
urllib3
watchdog
xlrd
parameterized
pika
pandas
numpy
python-dateutil
psycopg2
psutil